Integrate translation logic into watcher
